Extracted interface for events and refactored accordingly
